Change variables and functions names in subject_reader.py which makes variables and functions names more meaningful and understandable.
Change the index in line 14 to avoid the dependent on the constant in list_exercises.py.

Make a slight adjustment to the layout to make it look more beautiful in lists_warmup.py.

Use the if statement to check if it is less than 0 to simplify the code in quick_picks.py.

Correct grammatical errors in total_income.py